wSIR: Weighted Sliced Inverse Regression (wSIR) for supervised dimension reduction of spatial transcriptomics and single cell gene expression data

Weighted Sliced Inverse Regression (wSIR) is a supervised dimension reduction algorithm for spatial transcriptomics gene expression data. For a provided gene expression matrix and dataframe of each cell's spatial coordinates, wSIR creates a low-dimensional representation of the gene expression data that preserves the ability to predict spatial coordinates that was present in the gene expression data. Furthermore, wSIR provides interpretable loadings which allow for projection of new single-cell gene expression data into a low-dimensional space which preserves the spatial information present in the gene expression data.

See also

Author

Maintainer: Max Woollard mwoo5086@uni.sydney.edu.au (ORCID)

Authors:

Linh Nghiem (ORCID)

Shila Ghazanfar shazanfar@gmail.com (ORCID)

Other contributors:

Pratibha Panwar (ORCID) [contributor]